testing out paru-bin

+++ steps/01-base/01-install/setup.sh
@@ -6,12 +6,12 @@ SETUP_PATH=$PWD
 # TODO suppress type error
 if ! type "paru" > /dev/null; then
   sudo pacman --sync --needed --noconfirm base-devel
-  [[ -d /tmp/paru ]] && sudo rm --recursive /tmp/paru
-  git clone https://aur.archlinux.org/paru.git /tmp/paru
-  cd /tmp/paru
+  [[ -d /tmp/paru-bin ]] && sudo rm --recursive /tmp/paru-bin
+  git clone https://aur.archlinux.org/paru-bin.git /tmp/paru-bin
+  cd /tmp/paru-bin
   makepkg --syncdeps --install --needed --noconfirm
   cd $SETUP_PATH
-  sudo rm --recursive /tmp/paru
+  sudo rm --recursive /tmp/paru-bin
 fi
 
 paru -S --color=auto --needed --noconfirm $(grep "^[^#]" $SETUP_PATH/packages)
